requirements.txt
alembic==1.15.2
aniso8601==10.0.1
attrs==25.3.0
blinker==1.9.0
click==8.2.0
Flask==3.1.1
flask-cors==5.0.1
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2==2.9.10
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
rpds-py==0.24.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
